requirements.txt
SpeechRecognition
pydub
pytube
transformers
python-dotenv
# ffmpeg --> instalar no SO